Updating tests
